make TypedColumn serializable
