+++ src/main/scala/com/github/codelionx/dodo/types/TypedColumn.scala
@@ -2,35 +2,33 @@ package com.github.codelionx.dodo.types
 
 import java.util.Objects
 
+import com.github.codelionx.dodo.types.TypedColumn.BuilderAdapter
+
 import scala.collection.mutable
-import scala.reflect.ClassTag
 
 
-/**
-  * Represents a column of a dataset associated with a specific type. The cell data is stored in the correct
-  * (primitive) type.
-  *
-  * @tparam T the data type
-  */
-abstract class TypedColumnBase[T <: Any](implicit ev: ClassTag[T]) {
-
-  protected val tag: ClassTag[T] = ev
+object TypedColumn {
 
   /**
-    * Returns the [[com.github.codelionx.dodo.types.DataType]] associated with this column.
+    * Adapter class to bridge between our buffer-based [[com.github.codelionx.dodo.types.TypedColumnBuilder]] and the
+    * [[scala.collection.mutable.Builder]] interface
     */
-  def dataType: DataType[T]
+  private final class BuilderAdapter[T](ref: TypedColumn[T]) extends mutable.ReusableBuilder[T, TypedColumn[T]] {
 
-  /**
-    * Returns the backing array of this `TypedColumn`.
-    */
-  def array: Array[T]
+    private val internalBuilder = TypedColumnBuilder[T](ref.dataType, ref.name)(ref.tag)
 
-  /**
-    * Column name
-    */
-  def name: String
+    override def clear(): Unit = internalBuilder.clear()
+
+    override def result(): TypedColumn[T] = internalBuilder.toTypedColumn
+
+    override def +=(elem: T): BuilderAdapter.this.type = {
+      internalBuilder += elem
+      this
+    }
+  }
+  
 }
+
 
 /**
   * Represents a column of a dataset associated with a specific type that provides a rich interface to manipulate and
@@ -42,9 +40,10 @@ trait TypedColumn[T <: Any]
   extends TypedColumnBase[T]
     with TypedColumnArrayLike[T]
     with TypedColumnSeqLike[T, TypedColumn[T]]
-    with TypedColumnSorting[T, TypedColumn[T]] {
+    with TypedColumnSorting[T, TypedColumn[T]]
+    with Serializable {
 
-  override protected def newBuilder: mutable.Builder[T, TypedColumn[T]] = new BuilderAdapter
+  override protected def newBuilder: mutable.Builder[T, TypedColumn[T]] = new BuilderAdapter(this)
 
 
   // overrides of [[java.lang.Object]]
@@ -63,23 +62,5 @@ trait TypedColumn[T <: Any]
         |-------------------------------
         |${mkString(",")}
         |""".stripMargin
-
-  /**
-    * Adapter class to bridge between our buffer-based [[com.github.codelionx.dodo.types.TypedColumnBuilder]] and the
-    * [[scala.collection.mutable.Builder]] interface
-    */
-  private final class BuilderAdapter extends mutable.ReusableBuilder[T, TypedColumn[T]] {
-
-    private val internalBuilder = TypedColumnBuilder[T](dataType, name)(tag)
-
-    override def clear(): Unit = internalBuilder.clear()
-
-    override def result(): TypedColumn[T] = internalBuilder.toTypedColumn
-
-    override def +=(elem: T): BuilderAdapter.this.type = {
-      internalBuilder += elem
-      this
-    }
-  }
 
 }
